split class related visitors to separate files
