+++ src/index.ts
@@ -4,7 +4,9 @@ import { OpaqueType } from './visitors/opaque_type';
 import { TypeAnnotation } from './visitors/type_annotation';
 import { TypeCastExpression } from './visitors/type_cast_expression';
 import { TypeParameterDeclaration } from './visitors/type_parameter_declaration';
-import { ClassMethod, ClassDeclaration, DeclareClass } from './visitors/class_declaration';
+import { ClassMethod } from './visitors/class_method';
+import { ClassDeclaration } from './visitors/class_declaration';
+import { DeclareClass } from './visitors/declare_class';
 import { ExportNamedDeclaration, ExportDefaultDeclaration } from './visitors/export_declaration';
 import { InterfaceDeclaration } from './visitors/interface_declaration';
 import { OptionalMemberExpression } from './visitors/optional_member_expression';
