+++ src/visitors/class_method.ts
@@ -1,0 +1,8 @@
+import { ClassMethod } from '@babel/types';
+import { NodePath } from '@babel/traverse';
+
+export function ClassMethod(path: NodePath<ClassMethod>) {
+  if (path.node.kind === 'constructor') {
+    path.get('returnType').remove();
+  }
+}
